# build/build/resources/main/openapi.yaml
openapi: 3.0.2
info:
  title: Video Storage Server API
  version: "1.0"
servers:
- url: http://localhost:8080/v1
paths:
  /health:
    get:
      description: Return the health of the service as HTTP 200 status. Useful to
        check if everything is configured correctly.
      responses:
        "200":
          description: OK
  /files/{fileid}:
    delete:
      description: Delete a video file
      parameters:
      - explode: false
        in: path
        name: fileid
        required: true
        schema:
          type: string
        style: simple
      responses:
        "204":
          description: File was successfully removed
        "404":
          description: File not found
    get:
      description: Download a video file by fileid. The file name will be restored
        as it was when you uploaded it.
      parameters:
      - explode: false
        in: path
        name: fileid
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            video/mp4:
              schema:
                format: binary
                type: string
            video/mpeg:
              schema:
                format: binary
                type: string
          description: OK
          headers:
            Content-Disposition:
              explode: false
              schema:
                type: string
              style: simple
        "404":
          description: File not found
  /files:
    get:
      description: List uploaded files
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/UploadedFile'
                type: array
          description: File list
    post:
      description: Upload a video file
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/_files_get_request'
      responses:
        "201":
          description: File uploaded
          headers:
            Location:
              description: Created file location
              explode: false
              schema:
                type: string
              style: simple
        "400":
          description: Bad request
        "409":
          description: File exists
        "415":
          description: Unsupported Media Type
components:
  schemas:
    UploadedFile:
      example:
        size: 0
        name: name
        created_at: 2000-01-23T04:56:07.000+00:00
        fileid: fileid
      properties:
        fileid:
          type: string
        name:
          description: filename
          type: string
        size:
          description: file size (bytes)
          type: integer
        created_at:
          description: Time when the data was saved on the server side.
          format: date-time
          type: string
      required:
      - created_at
      - fileid
      - name
      - size
    _files_get_request:
      properties:
        data:
          format: binary
          type: string
      type: object